feat: 허브 간 이동 경로 구현

+++ core/src/main/java/com/sparta/core/dto/CompanyRequest.java
@@ -5,6 +5,7 @@ import jakarta.persistence.Enumerated;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record CompanyRequest(
@@ -20,10 +21,10 @@ public record CompanyRequest(
     CompanyTypeEnum companyType,
 
     @NotNull(message = "위도를 입력해주세요.")
-    Long latitude,
+    BigDecimal latitude,
 
     @NotNull(message = "경도를 입력해주세요.")
-    Long longitude,
+    BigDecimal longitude,
 
     @NotNull(message = "허브 아이디를 입력해주세요.")
     UUID hubId
